+++ libc/src/fcntl/posix_fadvise.cc
@@ -5,7 +5,8 @@
 
 int posix_fadvise(int fd, off_t offset, off_t len, int advice)
 {
-    long status = syscall4(fd, offset, len, advice, SYS_fadvise64);
+    long status = syscall4(fd, offset, len,
+                           unsigned(advice), SYS_fadvise64);
 
     if (status == 0)
         return status;
