+++ libc/src/unistd/chdir.cc
@@ -6,7 +6,7 @@
 
 int chdir(char const *path)
 {
-    long status = syscall1(long(path), SYS_chdir);
+    long status = syscall1(uintptr_t(path), SYS_chdir);
 
     if (status >= 0)
         return status;
