+++ libc/src/unistd/link.cc
@@ -6,7 +6,7 @@
 
 int link(char const *target, char const *link)
 {
-    long status = syscall2(long(target), long(link), SYS_link);
+    long status = syscall2(uintptr_t(target), uintptr_t(link), SYS_link);
 
     if (status >= 0)
         return status;
