+++ libc/src/unistd/lseek.cc
@@ -6,7 +6,7 @@
 
 off_t lseek(int fd, off_t off, int whence)
 {
-    long status = syscall3(long(fd), off, long(whence), SYS_lseek);
+    long status = syscall3(fd, off, unsigned(whence), SYS_lseek);
 
     if (status >= 0)
         return status;
