+++ libc/src/unistd/pread.cc
@@ -6,7 +6,7 @@
 
 ssize_t pread(int fd, void *buf, size_t sz, off_t ofs)
 {
-    long status = syscall4(fd, long(buf), sz, ofs, SYS_pread64);
+    long status = syscall4(fd, uintptr_t(buf), sz, ofs, SYS_pread64);
 
     if (status >= 0)
         return status;
