+++ libc/src/unistd/rmdir.cc
@@ -6,7 +6,7 @@
 
 int rmdir(char const *path)
 {
-    long status = syscall1(long(path), SYS_rmdir);
+    long status = syscall1(uintptr_t(path), SYS_rmdir);
 
     if (status >= 0)
         return status;
