+++ libc/src/unistd/symlink.cc
@@ -6,7 +6,7 @@
 
 int symlink(char const *target, char const *link)
 {
-    long status = syscall2(long(target), long(link), SYS_symlink);
+    long status = syscall2(uintptr_t(target), uintptr_t(link), SYS_symlink);
 
     if (status >= 0)
         return status;
