+++ libc/src/unistd/truncate.cc
@@ -6,7 +6,7 @@
 
 int truncate(char const *path, off_t sz)
 {
-    long status = syscall2(long(path), sz, SYS_truncate);
+    long status = syscall2(uintptr_t(path), sz, SYS_truncate);
 
     if (status >= 0)
         return status;
